removes the last dependencies on lttp in tests
