+++ test/TestBase.py
@@ -1,3 +1,3 @@
-from .bases import TestBase, WorldTestBase
+from .bases import WorldTestBase
 from warnings import warn
 warn("TestBase was renamed to bases", DeprecationWarning)
